requirements.txt
selenium==4.16.0
webdriver_manager==4.0.1
python-dotenv==1.0.0
loguru==0.7.2
PyQt6==6.6.1